moved the bitop code from RedisDataCommands to NullRedisString
