+++ geode-redis/src/main/java/org/apache/geode/redis/internal/data/NullRedisString.java
@@ -16,7 +16,13 @@
 
 package org.apache.geode.redis.internal.data;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.apache.geode.cache.Region;
+import org.apache.geode.redis.internal.executor.StripedExecutor;
+import org.apache.geode.redis.internal.executor.string.RedisStringCommands;
+import org.apache.geode.redis.internal.executor.string.RedisStringCommandsFunctionExecutor;
 import org.apache.geode.redis.internal.netty.Coder;
 
 /**
@@ -147,4 +153,161 @@ public class NullRedisString extends RedisString {
     }
     return newBytes.length;
   }
+
+
+  /**
+   * bitop is currently only implemented here. It does not have an implementation on
+   * RedisString which is a bit odd. This implementation only has a couple of places
+   * that care if a RedisString for "key" exists.
+   */
+  public int bitop(RedisDataCommands redisDataCommands,
+      String operation,
+      ByteArrayWrapper key, List<ByteArrayWrapper> sources) {
+    List<ByteArrayWrapper> sourceValues = new ArrayList<>();
+    int selfIndex = -1;
+    // Read all the source values, except for self, before locking the stripe.
+    RedisStringCommands commander =
+        new RedisStringCommandsFunctionExecutor(redisDataCommands.getRegion());
+    for (ByteArrayWrapper sourceKey : sources) {
+      if (sourceKey.equals(key)) {
+        // get self later after the stripe is locked
+        selfIndex = sourceValues.size();
+        sourceValues.add(null);
+      } else {
+        sourceValues.add(commander.get(sourceKey));
+      }
+    }
+    int indexOfSelf = selfIndex;
+    StripedExecutor stripedExecutor = redisDataCommands.getStripedExecutor();
+    return stripedExecutor.execute(key,
+        () -> doBitOp(redisDataCommands, operation, key, indexOfSelf, sourceValues));
+  }
+
+  private int doBitOp(RedisDataCommands redisDataCommands,
+      String operation,
+      ByteArrayWrapper key,
+      int selfIndex,
+      List<ByteArrayWrapper> sourceValues) {
+    if (selfIndex != -1) {
+      RedisString redisString = redisDataCommands.getRedisString(key);
+      if (redisString != null) {
+        sourceValues.set(selfIndex, redisString.getValue());
+      }
+    }
+    int maxLength = 0;
+    for (ByteArrayWrapper sourceValue : sourceValues) {
+      if (sourceValue != null && maxLength < sourceValue.length()) {
+        maxLength = sourceValue.length();
+      }
+    }
+    ByteArrayWrapper newValue;
+    switch (operation) {
+      case "AND":
+        newValue = and(sourceValues, maxLength);
+        break;
+      case "OR":
+        newValue = or(sourceValues, maxLength);
+        break;
+      case "XOR":
+        newValue = xor(sourceValues, maxLength);
+        break;
+      default: // NOT
+        newValue = not(sourceValues.get(0), maxLength);
+        break;
+    }
+    if (newValue.length() == 0) {
+      redisDataCommands.getRegion().remove(key);
+    } else {
+      RedisString redisString = redisDataCommands.getRedisStringForSet(key);
+      if (redisString == null) {
+        redisString = new RedisString(newValue);
+      } else {
+        redisString.set(newValue);
+      }
+      redisDataCommands.getRegion().put(key, redisString);
+    }
+    return newValue.length();
+  }
+
+  private ByteArrayWrapper and(List<ByteArrayWrapper> sourceValues, int max) {
+    byte[] dest = new byte[max];
+    for (int i = 0; i < max; i++) {
+      byte b = 0;
+      boolean firstByte = true;
+      for (ByteArrayWrapper sourceValue : sourceValues) {
+        byte sourceByte = 0;
+        if (sourceValue != null && i < sourceValue.length()) {
+          sourceByte = sourceValue.toBytes()[i];
+        }
+        if (firstByte) {
+          b = sourceByte;
+          firstByte = false;
+        } else {
+          b &= sourceByte;
+        }
+      }
+      dest[i] = b;
+    }
+    return new ByteArrayWrapper(dest);
+  }
+
+  private ByteArrayWrapper or(List<ByteArrayWrapper> sourceValues, int max) {
+    byte[] dest = new byte[max];
+    for (int i = 0; i < max; i++) {
+      byte b = 0;
+      boolean firstByte = true;
+      for (ByteArrayWrapper sourceValue : sourceValues) {
+        byte sourceByte = 0;
+        if (sourceValue != null && i < sourceValue.length()) {
+          sourceByte = sourceValue.toBytes()[i];
+        }
+        if (firstByte) {
+          b = sourceByte;
+          firstByte = false;
+        } else {
+          b |= sourceByte;
+        }
+      }
+      dest[i] = b;
+    }
+    return new ByteArrayWrapper(dest);
+  }
+
+  private ByteArrayWrapper xor(List<ByteArrayWrapper> sourceValues, int max) {
+    byte[] dest = new byte[max];
+    for (int i = 0; i < max; i++) {
+      byte b = 0;
+      boolean firstByte = true;
+      for (ByteArrayWrapper sourceValue : sourceValues) {
+        byte sourceByte = 0;
+        if (sourceValue != null && i < sourceValue.length()) {
+          sourceByte = sourceValue.toBytes()[i];
+        }
+        if (firstByte) {
+          b = sourceByte;
+          firstByte = false;
+        } else {
+          b ^= sourceByte;
+        }
+      }
+      dest[i] = b;
+    }
+    return new ByteArrayWrapper(dest);
+  }
+
+  private ByteArrayWrapper not(ByteArrayWrapper sourceValue, int max) {
+    byte[] dest = new byte[max];
+    if (sourceValue == null) {
+      for (int i = 0; i < max; i++) {
+        dest[i] = ~0;
+      }
+    } else {
+      byte[] cA = sourceValue.toBytes();
+      for (int i = 0; i < max; i++) {
+        dest[i] = (byte) (~cA[i] & 0xFF);
+      }
+    }
+    return new ByteArrayWrapper(dest);
+  }
+
 }
